LayoutTests/imported/w3c/web-platform-tests/custom-elements/state/state-css-selector-nth-of-expected.txt

First Element
First Sibling

Second Element
Second Sibling


PASS state selector has influence on nth-of when state is applied
PASS state selector only applies on given ident
PASS style is invalided on clear()
